LPD-62959 Move to tests folder
