+++ modules/test/playwright/tests/layout-content-page-editor-web/main/tests/editables.spec.ts
@@ -5,20 +5,20 @@
 
 import {expect, mergeTests} from '@playwright/test';
 
-import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
-import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
-import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../../fixtures/loginTest';
-import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
-import {pageManagementSiteTest} from '../../../fixtures/pageManagementSiteTest';
-import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
-import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
-import dragAndDropElement from '../../../utils/dragAndDropElement';
-import getRandomString from '../../../utils/getRandomString';
-import getBasicWebContentStructureId from '../../../utils/structured-content/getBasicWebContentStructureId';
-import chooseFileFromDocumentLibrary from './utils/chooseFileFromDocumentLibrary';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getPageDefinition from './utils/getPageDefinition';
+import {apiHelpersTest} from '../../../../fixtures/apiHelpersTest';
+import {featureFlagsTest} from '../../../../fixtures/featureFlagsTest';
+import {isolatedSiteTest} from '../../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../../fixtures/loginTest';
+import {pageEditorPagesTest} from '../../../../fixtures/pageEditorPagesTest';
+import {pageManagementSiteTest} from '../../../../fixtures/pageManagementSiteTest';
+import {clickAndExpectToBeHidden} from '../../../../utils/clickAndExpectToBeHidden';
+import {clickAndExpectToBeVisible} from '../../../../utils/clickAndExpectToBeVisible';
+import dragAndDropElement from '../../../../utils/dragAndDropElement';
+import getRandomString from '../../../../utils/getRandomString';
+import getBasicWebContentStructureId from '../../../../utils/structured-content/getBasicWebContentStructureId';
+import chooseFileFromDocumentLibrary from '../utils/chooseFileFromDocumentLibrary';
+import getFragmentDefinition from '../utils/getFragmentDefinition';
+import getPageDefinition from '../utils/getPageDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
