+++ modules/test/playwright/tests/layout-content-page-editor-web/main/tests/formContainer.spec.ts
@@ -12,33 +12,33 @@ import {
 import {Locator, Page, expect, mergeTests} from '@playwright/test';
 import path from 'path';
 
-import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
-import {dataApiHelpersTest} from '../../../fixtures/dataApiHelpersTest';
-import {displayPageTemplatesPagesTest} from '../../../fixtures/displayPageTemplatesPagesTest';
-import {documentLibraryPagesTest} from '../../../fixtures/documentLibraryPages.fixtures';
-import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
-import {fragmentsPagesTest} from '../../../fixtures/fragmentPagesTest';
-import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../../fixtures/loginTest';
-import {masterPagesPagesTest} from '../../../fixtures/masterPagesPagesTest';
-import {objectPagesTest} from '../../../fixtures/objectPagesTest';
-import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
-import {pageManagementSiteTest} from '../../../fixtures/pageManagementSiteTest';
-import {PageEditorPage} from '../../../pages/layout-content-page-editor-web/PageEditorPage';
-import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
-import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
-import {expandSection} from '../../../utils/expandSection';
-import fillAndClickOutside from '../../../utils/fillAndClickOutside';
-import getRandomString from '../../../utils/getRandomString';
-import {waitForAlert} from '../../../utils/waitForAlert';
-import {getObjectERC} from '../../setup/page-management-site/main/utils/getObjectERC';
-import {goToObjectEntity} from '../../setup/page-management-site/main/utils/goToObjectEntity';
-import chooseFileFromDocumentLibrary from './utils/chooseFileFromDocumentLibrary';
-import getContainerDefinition from './utils/getContainerDefinition';
-import getFormContainerDefinition from './utils/getFormContainerDefinition';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getPageDefinition from './utils/getPageDefinition';
-import getWidgetDefinition from './utils/getWidgetDefinition';
+import {apiHelpersTest} from '../../../../fixtures/apiHelpersTest';
+import {dataApiHelpersTest} from '../../../../fixtures/dataApiHelpersTest';
+import {displayPageTemplatesPagesTest} from '../../../../fixtures/displayPageTemplatesPagesTest';
+import {documentLibraryPagesTest} from '../../../../fixtures/documentLibraryPages.fixtures';
+import {featureFlagsTest} from '../../../../fixtures/featureFlagsTest';
+import {fragmentsPagesTest} from '../../../../fixtures/fragmentPagesTest';
+import {isolatedSiteTest} from '../../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../../fixtures/loginTest';
+import {masterPagesPagesTest} from '../../../../fixtures/masterPagesPagesTest';
+import {objectPagesTest} from '../../../../fixtures/objectPagesTest';
+import {pageEditorPagesTest} from '../../../../fixtures/pageEditorPagesTest';
+import {pageManagementSiteTest} from '../../../../fixtures/pageManagementSiteTest';
+import {PageEditorPage} from '../../../../pages/layout-content-page-editor-web/PageEditorPage';
+import {clickAndExpectToBeHidden} from '../../../../utils/clickAndExpectToBeHidden';
+import {clickAndExpectToBeVisible} from '../../../../utils/clickAndExpectToBeVisible';
+import {expandSection} from '../../../../utils/expandSection';
+import fillAndClickOutside from '../../../../utils/fillAndClickOutside';
+import getRandomString from '../../../../utils/getRandomString';
+import {waitForAlert} from '../../../../utils/waitForAlert';
+import {getObjectERC} from '../../../setup/page-management-site/main/utils/getObjectERC';
+import {goToObjectEntity} from '../../../setup/page-management-site/main/utils/goToObjectEntity';
+import chooseFileFromDocumentLibrary from '../utils/chooseFileFromDocumentLibrary';
+import getContainerDefinition from '../utils/getContainerDefinition';
+import getFormContainerDefinition from '../utils/getFormContainerDefinition';
+import getFragmentDefinition from '../utils/getFragmentDefinition';
+import getPageDefinition from '../utils/getPageDefinition';
+import getWidgetDefinition from '../utils/getWidgetDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
