+++ modules/test/playwright/tests/layout-content-page-editor-web/main/tests/fragmentComposition.spec.ts
@@ -7,18 +7,18 @@ import {expect, mergeTests} from '@playwright/test';
 import {createReadStream} from 'fs';
 import path from 'path';
 
-import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
-import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
-import {fragmentsPagesTest} from '../../../fixtures/fragmentPagesTest';
-import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../../fixtures/loginTest';
-import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
-import getRandomString from '../../../utils/getRandomString';
-import getBasicWebContentStructureId from '../../../utils/structured-content/getBasicWebContentStructureId';
-import {waitForAlert} from '../../../utils/waitForAlert';
-import getContainerDefinition from './utils/getContainerDefinition';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getPageDefinition from './utils/getPageDefinition';
+import {apiHelpersTest} from '../../../../fixtures/apiHelpersTest';
+import {featureFlagsTest} from '../../../../fixtures/featureFlagsTest';
+import {fragmentsPagesTest} from '../../../../fixtures/fragmentPagesTest';
+import {isolatedSiteTest} from '../../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../../fixtures/loginTest';
+import {pageEditorPagesTest} from '../../../../fixtures/pageEditorPagesTest';
+import getRandomString from '../../../../utils/getRandomString';
+import getBasicWebContentStructureId from '../../../../utils/structured-content/getBasicWebContentStructureId';
+import {waitForAlert} from '../../../../utils/waitForAlert';
+import getContainerDefinition from '../utils/getContainerDefinition';
+import getFragmentDefinition from '../utils/getFragmentDefinition';
+import getPageDefinition from '../utils/getPageDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
