+++ modules/test/playwright/tests/layout-content-page-editor-web/main/tests/fragments.spec.ts
@@ -6,41 +6,41 @@
 import {ObjectDefinitionAPI} from '@liferay/object-admin-rest-client-js';
 import {Page, expect, mergeTests} from '@playwright/test';
 
-import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
-import {displayPageTemplatesPagesTest} from '../../../fixtures/displayPageTemplatesPagesTest';
-import {documentLibraryPagesTest} from '../../../fixtures/documentLibraryPages.fixtures';
-import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
-import {fragmentsPagesTest} from '../../../fixtures/fragmentPagesTest';
-import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../../fixtures/loginTest';
-import {masterPagesPagesTest} from '../../../fixtures/masterPagesPagesTest';
-import {objectPagesTest} from '../../../fixtures/objectPagesTest';
-import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
-import {pageManagementSiteTest} from '../../../fixtures/pageManagementSiteTest';
-import {PageEditorPage} from '../../../pages/layout-content-page-editor-web/PageEditorPage';
-import {checkAccessibility} from '../../../utils/checkAccessibility';
-import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
-import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
-import dragAndDropElement from '../../../utils/dragAndDropElement';
-import getGlobalSiteId from '../../../utils/getGlobalSiteId';
-import getRandomString from '../../../utils/getRandomString';
-import {hoverAndExpectToBeVisible} from '../../../utils/hoverAndExpectToBeVisible';
-import {performLogout} from '../../../utils/performLogin';
+import {apiHelpersTest} from '../../../../fixtures/apiHelpersTest';
+import {displayPageTemplatesPagesTest} from '../../../../fixtures/displayPageTemplatesPagesTest';
+import {documentLibraryPagesTest} from '../../../../fixtures/documentLibraryPages.fixtures';
+import {featureFlagsTest} from '../../../../fixtures/featureFlagsTest';
+import {fragmentsPagesTest} from '../../../../fixtures/fragmentPagesTest';
+import {isolatedSiteTest} from '../../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../../fixtures/loginTest';
+import {masterPagesPagesTest} from '../../../../fixtures/masterPagesPagesTest';
+import {objectPagesTest} from '../../../../fixtures/objectPagesTest';
+import {pageEditorPagesTest} from '../../../../fixtures/pageEditorPagesTest';
+import {pageManagementSiteTest} from '../../../../fixtures/pageManagementSiteTest';
+import {PageEditorPage} from '../../../../pages/layout-content-page-editor-web/PageEditorPage';
+import {checkAccessibility} from '../../../../utils/checkAccessibility';
+import {clickAndExpectToBeHidden} from '../../../../utils/clickAndExpectToBeHidden';
+import {clickAndExpectToBeVisible} from '../../../../utils/clickAndExpectToBeVisible';
+import dragAndDropElement from '../../../../utils/dragAndDropElement';
+import getGlobalSiteId from '../../../../utils/getGlobalSiteId';
+import getRandomString from '../../../../utils/getRandomString';
+import {hoverAndExpectToBeVisible} from '../../../../utils/hoverAndExpectToBeVisible';
+import {performLogout} from '../../../../utils/performLogin';
 import getBasicWebContentStructureId, {
 	getWebContentStructureId,
-} from '../../../utils/structured-content/getBasicWebContentStructureId';
-import {waitForAlert} from '../../../utils/waitForAlert';
-import {journalPagesTest} from '../../journal-web/main/fixtures/journalPagesTest';
+} from '../../../../utils/structured-content/getBasicWebContentStructureId';
+import {waitForAlert} from '../../../../utils/waitForAlert';
+import {journalPagesTest} from '../../../journal-web/main/fixtures/journalPagesTest';
 import {
 	ANIMAL_DDM_STRUCTURE_KEY,
 	ANIMAL_DDM_TEMPLATE_KEY,
-} from '../../setup/page-management-site/main/constants/animals';
-import {getObjectERC} from '../../setup/page-management-site/main/utils/getObjectERC';
-import {goToObjectEntity} from '../../setup/page-management-site/main/utils/goToObjectEntity';
-import getContainerDefinition from './utils/getContainerDefinition';
-import getFormContainerDefinition from './utils/getFormContainerDefinition';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getPageDefinition from './utils/getPageDefinition';
+} from '../../../setup/page-management-site/main/constants/animals';
+import {getObjectERC} from '../../../setup/page-management-site/main/utils/getObjectERC';
+import {goToObjectEntity} from '../../../setup/page-management-site/main/utils/goToObjectEntity';
+import getContainerDefinition from '../utils/getContainerDefinition';
+import getFormContainerDefinition from '../utils/getFormContainerDefinition';
+import getFragmentDefinition from '../utils/getFragmentDefinition';
+import getPageDefinition from '../utils/getPageDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
