+++ modules/test/playwright/tests/layout-content-page-editor-web/main/tests/grid.spec.ts
@@ -5,18 +5,18 @@
 
 import {expect, mergeTests} from '@playwright/test';
 
-import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
-import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
-import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../../fixtures/loginTest';
-import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
-import {pageManagementSiteTest} from '../../../fixtures/pageManagementSiteTest';
-import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
-import getRandomString from '../../../utils/getRandomString';
-import getContainerDefinition from './utils/getContainerDefinition';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getGridDefinition from './utils/getGridDefinition';
-import getPageDefinition from './utils/getPageDefinition';
+import {apiHelpersTest} from '../../../../fixtures/apiHelpersTest';
+import {featureFlagsTest} from '../../../../fixtures/featureFlagsTest';
+import {isolatedSiteTest} from '../../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../../fixtures/loginTest';
+import {pageEditorPagesTest} from '../../../../fixtures/pageEditorPagesTest';
+import {pageManagementSiteTest} from '../../../../fixtures/pageManagementSiteTest';
+import {clickAndExpectToBeHidden} from '../../../../utils/clickAndExpectToBeHidden';
+import getRandomString from '../../../../utils/getRandomString';
+import getContainerDefinition from '../utils/getContainerDefinition';
+import getFragmentDefinition from '../utils/getFragmentDefinition';
+import getGridDefinition from '../utils/getGridDefinition';
+import getPageDefinition from '../utils/getPageDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
