+++ modules/test/playwright/tests/layout-content-page-editor-web/main/tests/sidebar.spec.ts
@@ -8,36 +8,36 @@ import {expect, mergeTests} from '@playwright/test';
 import {createReadStream} from 'fs';
 import path from 'path';
 
-import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
-import {applicationsMenuPageTest} from '../../../fixtures/applicationsMenuPageTest';
-import {collectionsPagesTest} from '../../../fixtures/collectionsPagesTest';
-import {dataApiHelpersTest} from '../../../fixtures/dataApiHelpersTest';
-import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
-import {fragmentsPagesTest} from '../../../fixtures/fragmentPagesTest';
-import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../../fixtures/loginTest';
-import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
-import {pageManagementSiteTest} from '../../../fixtures/pageManagementSiteTest';
-import {pageViewModePagesTest} from '../../../fixtures/pageViewModePagesTest';
-import {checkAccessibility} from '../../../utils/checkAccessibility';
-import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
-import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
-import createUserWithPermissions from '../../../utils/createUserWithPermissions';
-import {expandSection} from '../../../utils/expandSection';
-import getRandomString from '../../../utils/getRandomString';
-import {hoverAndExpectToBeVisible} from '../../../utils/hoverAndExpectToBeVisible';
-import {performUserSwitch} from '../../../utils/performLogin';
-import {closeProductMenu, openProductMenu} from '../../../utils/productMenu';
-import getBasicWebContentStructureId from '../../../utils/structured-content/getBasicWebContentStructureId';
-import {waitForAlert} from '../../../utils/waitForAlert';
-import {journalPagesTest} from '../../journal-web/main/fixtures/journalPagesTest';
-import {ANIMALS_COLLECTION_NAME} from '../../setup/page-management-site/main/constants/animals';
-import {getObjectERC} from '../../setup/page-management-site/main/utils/getObjectERC';
-import getCollectionDefinition from './utils/getCollectionDefinition';
-import getFormContainerDefinition from './utils/getFormContainerDefinition';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getGridDefinition from './utils/getGridDefinition';
-import getPageDefinition from './utils/getPageDefinition';
+import {apiHelpersTest} from '../../../../fixtures/apiHelpersTest';
+import {applicationsMenuPageTest} from '../../../../fixtures/applicationsMenuPageTest';
+import {collectionsPagesTest} from '../../../../fixtures/collectionsPagesTest';
+import {dataApiHelpersTest} from '../../../../fixtures/dataApiHelpersTest';
+import {featureFlagsTest} from '../../../../fixtures/featureFlagsTest';
+import {fragmentsPagesTest} from '../../../../fixtures/fragmentPagesTest';
+import {isolatedSiteTest} from '../../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../../fixtures/loginTest';
+import {pageEditorPagesTest} from '../../../../fixtures/pageEditorPagesTest';
+import {pageManagementSiteTest} from '../../../../fixtures/pageManagementSiteTest';
+import {pageViewModePagesTest} from '../../../../fixtures/pageViewModePagesTest';
+import {checkAccessibility} from '../../../../utils/checkAccessibility';
+import {clickAndExpectToBeHidden} from '../../../../utils/clickAndExpectToBeHidden';
+import {clickAndExpectToBeVisible} from '../../../../utils/clickAndExpectToBeVisible';
+import createUserWithPermissions from '../../../../utils/createUserWithPermissions';
+import {expandSection} from '../../../../utils/expandSection';
+import getRandomString from '../../../../utils/getRandomString';
+import {hoverAndExpectToBeVisible} from '../../../../utils/hoverAndExpectToBeVisible';
+import {performUserSwitch} from '../../../../utils/performLogin';
+import {closeProductMenu, openProductMenu} from '../../../../utils/productMenu';
+import getBasicWebContentStructureId from '../../../../utils/structured-content/getBasicWebContentStructureId';
+import {waitForAlert} from '../../../../utils/waitForAlert';
+import {journalPagesTest} from '../../../journal-web/main/fixtures/journalPagesTest';
+import {ANIMALS_COLLECTION_NAME} from '../../../setup/page-management-site/main/constants/animals';
+import {getObjectERC} from '../../../setup/page-management-site/main/utils/getObjectERC';
+import getCollectionDefinition from '../utils/getCollectionDefinition';
+import getFormContainerDefinition from '../utils/getFormContainerDefinition';
+import getFragmentDefinition from '../utils/getFragmentDefinition';
+import getGridDefinition from '../utils/getGridDefinition';
+import getPageDefinition from '../utils/getPageDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
